+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 17.2.0 (upcoming)
+
+* Pending changelog
+
 ## 17.1.0 (upcoming)
 
 **New features:**